Reworking Connected Components API
